Remove default values for config_dir and data_dir
Both properties represent directories, so their default values should be
a path or `None`. Their previous default values ('false') are incorrect.
This hasn't caused an error because these valuse were only used to
generate rst and sample yml files.

Specifying `None` as a default value is redundant: it is currently read
in `config_manage.py` and is resovled to `None` if missing.