Refactor infrastructure_url/port values; fix tests
This edit is possible because schema default values are loaded
automatically.

Here's what the code did:
- web_port was read from kwargs, assigned null by default
- galaxy_infrastructure_url was read from kwargs, assigned null by default
- as a result, schema defaults were never used; however:
- if the url was not set in kwargs, it was rebuilt to the same value as
  the schema default minus the port, which was later specified in a
  different location.
- set an attribute indicating if the url was set from kwargs
- substitute 'HOST_IP' in url

Given that the schema defaults are now preloaded, most of this logic is
now redundant.

Updated logic:
- set an attribute indicating if the url was set from kwargs
- substitute 'HOST_IP' in url

I have removed the comments that duplicate the text in the schema.

Potential side effect: now galaxy_infrastructure_url always contains the
default port unless it is explicitly set to null. I don't expect this to
be an issue, but I am not completely sure.

Integration tests for both defaults should pass.