Correct _load_schema method signature; edit comments
Following the principle of least surprise.
The stub methods are not invoked because the parent's __init__() method
is not called. They can be safely removed. However, if a future commit adds
a call to parent's __init__, without these stub methods the parent's
methods will be called, which may result in a surprise (e.g., attributes
would be created for all schema properties).