Fix tests: tsi and gxy may have different stored versions
This is how the current codebase works.
If the database is combined, we only store the galaxy version in the
  migrate_version table
But if we have 2 separate databases, we store 2 different versions
(currently, 179 for galaxy and 17 for install).

This commit adjusts the tests to test for this case; it also fixes the
error that prevented an upgrade from proceeding with separate databases
and automigrate set.