Fix legacy get_chunk: use offset instead of chunk index
This was a bug waiting to happen:
The are 2 variations of the method signature on tabular datatypes: one
taking a chunk index, the other one taking an absolute offset.
The get_chunk() method is called from the legacy dataset controller;
thus, the argument could be interpreted either as an absolute offset or
a chunk index. This has not been a problem because the only call to this
method had 0 hardcoded as the argument, so it didn't matter how it was
interpreted.

Additionally, this modification removes a mypy error, so now we can add
type hints.