Ensure request-scoped SA session is always closed
In some cases we never reach the end of the handle_request() method; as
a result, the request-scoped SQLAlchemy session was never closed,
causing database transactions being stuck in idle state.

Here we set/unset the request_id on the scoped session within a
try/finally block, thus ensuring the session will be always closed.