Handle CompoundSelect statement
Calling union() on a statement creates a CompoundSelect object. Unlike a
Select object, a CompoundSelect requires a model to be explicitly
selected via `from_statement`. Otherwise, it will return tuples, not
models.

Checking the statement type is not a pretty solution. A better way to
do this is to break up the index_query method, which I'll leave for a
separate commit.